move qrep status page into mirrors page instead of redirecting
