+++ ui/app/mirrors/[mirrorId]/page.tsx
@@ -7,9 +7,10 @@ import { MirrorStatusResponse } from '@/grpc_generated/route';
 import { Header } from '@/lib/Header';
 import { LayoutMain } from '@/lib/Layout';
 import { GetFlowHttpAddressFromEnv } from '@/rpc/http';
-import { redirect } from 'next/navigation';
 import { CDCMirror } from './cdc';
 import NoMirror from './nomirror';
+import QrepGraph from './qrepGraph';
+import QRepStatusTable, { QRepPartitionStatus } from './qrepStatusTable';
 import SyncStatus from './syncStatus';
 
 type EditMirrorProps = {
@@ -30,6 +31,17 @@ async function getMirrorStatus(mirrorId: string) {
   return json;
 }
 
+function setFlowState(flowJobName: string, requestedFlowState: FlowStatus) {
+  return fetch(`/api/mirrors/state_change`, {
+    method: 'POST',
+    body: JSON.stringify({
+      flowJobName,
+      requestedFlowState,
+    }),
+    cache: 'no-store',
+  });
+}
+
 export default async function ViewMirror({
   params: { mirrorId },
 }: EditMirrorProps) {
@@ -38,7 +50,7 @@ export default async function ViewMirror({
     return <div>No mirror status found!</div>;
   }
 
-  let mirrorInfo = await prisma.flows.findFirst({
+  const mirrorInfo = await prisma.flows.findFirst({
     select: {
       created_at: true,
       workflow_id: true,
@@ -49,7 +61,7 @@ export default async function ViewMirror({
     },
   });
 
-  let syncs = await prisma.cdc_batches.findMany({
+  const syncs = await prisma.cdc_batches.findMany({
     where: {
       flow_name: mirrorId,
       start_time: {
@@ -113,29 +125,85 @@ export default async function ViewMirror({
         isNotPaused={isNotPaused}
       />
     );
-  } else {
-    redirect(`/mirrors/status/qrep/${mirrorId}`);
-  }
 
-  return (
-    <LayoutMain alignSelf='flex-start' justifySelf='flex-start' width='full'>
-      <div
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'space-between',
-          paddingRight: '2rem',
-        }}
-      >
-        <Header variant='title2'>{mirrorId}</Header>
-        {actionsDropdown}
-      </div>
-      <CDCMirror
-        rows={rows}
-        createdAt={mirrorInfo?.created_at}
-        syncStatusChild={syncStatusChild}
-        status={mirrorStatus}
-      />
-    </LayoutMain>
-  );
+    return (
+      <LayoutMain alignSelf='flex-start' justifySelf='flex-start' width='full'>
+        <div
+          style={{
+            display: 'flex',
+            alignItems: 'center',
+            justifyContent: 'space-between',
+            paddingRight: '2rem',
+          }}
+        >
+          <Header variant='title2'>{mirrorId}</Header>
+          {actionsDropdown}
+        </div>
+        <CDCMirror
+          rows={rows}
+          createdAt={mirrorInfo?.created_at}
+          syncStatusChild={syncStatusChild}
+          status={mirrorStatus}
+        />
+      </LayoutMain>
+    );
+  } else {
+    const runs = await prisma.qrep_partitions.findMany({
+      where: {
+        flow_name: mirrorId,
+        start_time: {
+          not: null,
+        },
+        rows_in_partition: {
+          not: 0,
+        },
+      },
+      orderBy: {
+        start_time: 'desc',
+      },
+    });
+
+    const partitions: QRepPartitionStatus[] = runs.map((run) => ({
+      partitionId: run.partition_uuid,
+      startTime: run.start_time,
+      endTime: run.end_time,
+      pulledRows: run.rows_in_partition,
+      syncedRows: Number(run.rows_synced),
+    }));
+
+    return (
+      <LayoutMain alignSelf='flex-start' justifySelf='flex-start' width='full'>
+        <div
+          style={{
+            display: 'flex',
+            alignItems: 'center',
+            justifyContent: 'space-between',
+            paddingRight: '2rem',
+          }}
+        >
+          <Header variant='title2'>{mirrorId}</Header>
+          <input
+            type='button'
+            value='Pause'
+            onClick={() => setFlowState(mirrorId, FlowStatus.STATUS_PAUSED)}
+          />
+          <input
+            type='button'
+            value='Resume'
+            onClick={() => setFlowState(mirrorId, FlowStatus.STATUS_RUNNING)}
+          />
+        </div>
+        <QrepGraph
+          syncs={partitions.map((partition) => ({
+            partitionID: partition.partitionId,
+            startTime: partition.startTime,
+            endTime: partition.endTime,
+            numRows: partition.pulledRows,
+          }))}
+        />
+        <br></br>
+        <QRepStatusTable flowJobName={mirrorId} partitions={partitions} />
+      </LayoutMain>
+    );
+  }
 }
