+++ ui/app/mirrors/[mirrorId]/qrepGraph.tsx
@@ -9,7 +9,7 @@ import { Label } from '@/lib/Label';
 import { BarChart } from '@tremor/react';
 import { useEffect, useState } from 'react';
 import ReactSelect from 'react-select';
-import aggregateCountsByInterval from '../../../[mirrorId]/aggregatedCountsByInterval';
+import aggregateCountsByInterval from './aggregatedCountsByInterval';
 
 type QrepStatusRow = {
   partitionID: string;
