+++ ui/app/mirrors/[mirrorId]/qrepStatusTable.tsx
